setuptools and pkg_resources build via PYOBJ.COM
It is possible setuptools and pkg_resources in site-packages, but for
experiment's sake I just wanted it to build correctly.

I found that both setuptools and pkg_resources use the same set of
vendored packages, so moved the _vendor folder outside to make it common
to both packages.

PYOBJ.COM complains if the import is not "obvious" ie if the package(s)
being imported are actually doing some hijinks with available names.

The example is in _vendor/six.py: it doesn't contain any submodules,
but acts as if it has a bunch of them via the "MovedModule" trick. I
don't expect PYOBJ.COM to find these (halting problem and whatnot), but
dynamic imports are pretty annoying.

Also: sandbox.py in setuptools was importing itself. I don't understand
why that has to be done.